+++ plugins/postcss-random-function/src/index.ts
@@ -9,6 +9,7 @@ export type pluginOptions = {
 };
 
 const FUNCTION_CALL_REGEX = /(?<![-\w])(?:random)\(/i;
+const IS_PROPERTY_REGEX = /^property$/i;
 
 const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 	const options: pluginOptions = Object.assign(
@@ -24,6 +25,10 @@ const creator: PluginCreator<pluginOptions> = (opts?: pluginOptions) => {
 		postcssPlugin: 'postcss-random-function',
 		Declaration(decl): void {
 			if (!FUNCTION_CALL_REGEX.test(decl.value)) {
+				return;
+			}
+
+			if (options.preserve && decl.parent?.type === 'atrule' && IS_PROPERTY_REGEX.test(decl.parent.name)) {
 				return;
 			}